+++ Tests/CSharp/MemberTests/EventMemberTests.cs
@@ -3,7 +3,7 @@ using ICSharpCode.CodeConverter.Tests.TestRunners;
 using ICSharpCode.CodeConverter.CSharp;
 using Xunit;
 
-namespace ICSharpCode.CodeConverter.Tests.CSharp
+namespace ICSharpCode.CodeConverter.Tests.CSharp.MemberTests
 {
     public class EventMemberTests : ConverterTestBase
     {