+++ Tests/TestData/MultiFileCharacterization/SourceFiles/VbLibrary/PartialClassPart.vb
@@ -2,4 +2,5 @@
 ''' Test that partial class modifiers are added to both parts
 ''' </summary>
 Partial Class AClass
+    Private anArrayWithNonStaticInitializerReferencingOtherPart(anInt) As Integer
 End Class