+++ Tests/TestData/MultiFileCharacterization/VBToCSResults/ConvertWholeSolution/VbLibrary/PartialClassPart.cs
@@ -6,5 +6,8 @@ namespace VbLibrary
 {
     internal partial class AClass
     {
+        static int[] initialAnArrayWithNonStaticInitializerReferencingOtherPart() => new int[anInt + 1];
+
+        private int[] anArrayWithNonStaticInitializerReferencingOtherPart;
     }
 }